chore: move src hosting from forgejo.kainnef.com to codeberg.org (#1)
Although I kinda want to run on a self-hosted Forgejo instance... I also don't want to open that instance up for public registration to allow people to collaborate/contribute to testtrim, and then have to deal with spam accounts that have the capability to create and push repos with arbitrary content onto my self-hosted instance.

So, I think it seems like a better idea to use codeberg for this purpose; still leveraging the Open Source Forgejo project, but taking a little bit of potential maintenance burden away from me.

Reviewed-on: https://codeberg.org/testtrim/testtrim/pulls/1
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>