in-process out-of-band performance tracing using `tracing` library (#17)
Should address #8 when complete.

Rather than providing performance/profiling data directly as part of return values from functions, this uses the `tracing` library to instrument relevant codepaths with a `perftrace` attribute, and then will use a custom subscriber to fetch that data "out-of-band" so that it can be flexibly changed without having to punch data through multiple layers of function calls.

Reviewed-on: https://codeberg.org/testtrim/testtrim/pulls/17
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>