fix: don't skip relevant instrumentation points (#23)
Fixes #22.

Discovered that instrumentation points, which can have multiple files associated with them in the coverage map, might have both internal project files and external crate files listed.  It isn't clear to me how to reproduce this in a test project, but an analysis of the case where it was observed made it clear that the internal file was the "more true" of the cases (as it contained the function being executed).  So the shortcut to remove the external dependency tracking needs to "prefer" treating a file as internal if it is within the target project.

Reviewed-on: https://codeberg.org/testtrim/testtrim/pulls/23
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>