fix: coverage data mix-and-match between executables must be prevented due to identifier conflicts (#24)
While investigating #22, the primary reason for the problem seemed to be the change in #20 where coverage data was put into a map, rather than doing a linear search of all executables.  The map was having conflicts during initialization that were being silently ignored, as multiple binaries might have the same name_hash and fn_hash in their coverage map.

To fix this, the coverage maps are kept separate for each test binary now, and a guard is in-place to prevent duplicate values.

This means that the test counts in simulate-history (#22) should now be correct, but won't match either the original implementation or the performance optimized implementation. 🤷

Reviewed-on: https://codeberg.org/testtrim/testtrim/pulls/24
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>