feat: add an OCI container published for each release for testtrim run-server (#144)
This PR adds a published OCI (Docker) container for every release build which will run the server.  This is a reasonable prereq to having a shared server available for a CI system.  The container is rebuilt and pushed on every release tag.

Note: some small pieces of the container push are untested and need an end-to-end release test; some patches might be done shortly to fix issues if they exist.
Reviewed-on: https://codeberg.org/testtrim/testtrim/pulls/144
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>