ci: check-in .sqlx files, rather than always generating at build-time (#169)
This is necessary to allow `flake.nix` to be used remotely to provide the testtrim package, as the build environment in there doesn't provide PostgreSQL, run migrations, etc.  This is *probably* also necessary to create a binary that can be used with `cargo install`.  So between those things, it seems like it makes sense to check-in the `.sqlx` files and use CI to verify that they're updated.

Part of https://codeberg.org/testtrim/testtrim/issues/59, considering these reasons.

Testing:
- [x] Clean successful build
- [x] Verified that if `.sqlx` is out-of-date, build fails

Reviewed-on: https://codeberg.org/testtrim/testtrim/pulls/169
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>