chore: set deps to current specific versions (#173)
Originally I had the mindset of setting all deps to `"*"`, the latest, to make maintenance of dependencies easier.  Now with Renovate, this has the effect of causing all updates to get merged into the lockfile update, rather than being processed and inspected independently.  So this switches to being specific versions so that Renovate can do it's job correctly.

Reviewed-on: https://codeberg.org/testtrim/testtrim/pulls/173
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>