fix: temporary hacky workaround for Go DNS resolution (#223)
Per https://codeberg.org/testtrim/testtrim/issues/217 -- our strace doesn't work correctly for DNS tracing if the socket is connected on one thread, and then read on another thread. Until I can figure out a way to address that issue, this is a little hack that looks for the DNS "flags" response (the third & fourth bytes) and see if it looks like a typical recursive query response.

So in this case where we've performed a read but we don't know what the file descriptor is, but the two bytes match the response pattern for a query, then we're going to pretend that it is a response that we received from a DNS server.  This hack will continue in two other places; we'll be OK if we can't parse this as a query response (search: hack_query_response), and we'll remove the 0.0.0.0:53 connection trace.

This hack is valuable as it allows us to track connections to proxy.golang.org during testtrim's golang acceptance test, preventing it from rerunning indefinitely when testtrim runs on itself.

Reviewed-on: https://codeberg.org/testtrim/testtrim/pulls/223
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>