fix(rust): Coverage parsing overstating affected files (#227)
Rust LLVM coverage parsing assumed all filepaths were relevant to the code function; however, the data contained an index into the array which should have indicated a specific file that defined the function, and this implementation wasn't using that index.

This is the cause of the bug identified in https://codeberg.org/testtrim/testtrim/issues/212 -- tests being triggered that didn't make any sense for the code being modified.

Unfortunately this indicated that one test case similar to https://codeberg.org/testtrim/testtrim/issues/52 -- modification of the definition of a `lazy_static` -- was being falsely flagged as working when in reality it was subject to this bug.  A new note will be added to https://codeberg.org/testtrim/testtrim/issues/52 indicating that `lazy_static` is also broken.

Reviewed-on: https://codeberg.org/testtrim/testtrim/pulls/227
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>