test: Split integration tests into smaller chunks (#254)
The intent of this is to allow the longer tests to run in-parallel.  More overall work is done due to the need to re-checkout and re-compile for first commit of each test, but the interactive test wait time is shorter.  Theoretically there could be fewer tests targeted as well, especially if we implement function-level checking in the future.

Reviewed-on: https://codeberg.org/testtrim/testtrim/pulls/254
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>