refactor: organize testtrim into a Rust workspace; add PoC eBPF-based tracing (#257)
Currently this PR contains a proof-of-concept of an eBPF-based tracer with aya-rs, which is not implemented into testtrim yet.

I've decided that it's a good idea to merge it as-is because it reorganizes the testtrim project into a Rust workspace.  It will be easier to make smaller incremental improvements and add comprehensive tests once this base refactoring is merged in, without conflicting on other changes like @mathieu-bot's renovate PRs.

Reviewed-on: https://codeberg.org/testtrim/testtrim/pulls/257
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>