fix: syscall child syscalls can be emitted before "inherited" config is done (#294)
Ensures that the results of process-creation syscalls (Clone) are processed before any syscalls that occur from within those processes.

For example, we can get a series of syscalls like this:

- `25979 clone3([snip] <unfinished ...>`
- `26006 (some random syscall as process 26006 starts...)`
- `25979 <... clone3 resumed> => {parent_tid=[26006]}, 88) = 26006`

The parent process is starting a thread or subprocess, but there's no guarantee that the syscall with that PID is returned to the parent *before* the new child (26006) starts to execute.  This would prevent any syscalls from the child from being processed with the right context that should be inherited from the parent -- for example, its working directory.

Also includes a minor fix to `testtrim-syscall-test-app` to prevent zombie child processes.

Prerequisite to re-enabling testtrim in it's own CI; discovered while testing that.

Reviewed-on: https://codeberg.org/testtrim/testtrim/pulls/294
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>