ci: Fix intermittent freezes in testtrim due to presumed leftover processes (#305)
Possible cause of #304 -- when a process shuts down with threads via the `exit_group()` syscall, the strace tracer didn't recognize all the processes to close them in the Subscriber, which would prevent the DedicatedOperator from sending an EOF and closing the TraceClient.

Marking as "ci" -- this is only relevant for testtrim running on itself with nested strace.

Reviewed-on: https://codeberg.org/testtrim/testtrim/pulls/305
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>