build: update release API access to explicitly pass release body (#308)
When creating a release, we were previously just passing the tag and assuming codeberg would create a release with the body from the tag message.  With v0.9.0 this didn't work; the title and body were missing.  This could be an upstream regression in Forgejo since codeberg was upgraded to v10 today, or it could be a race condition in immediately creating the release right after uploading the tag.  Either way, we have the data in the release process to just populate it in the API call and prevent this problem.

Reviewed-on: https://codeberg.org/testtrim/testtrim/pulls/308
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>