fix: add timeouts for internal nested strace coordination (diagnostic) (#331)
This is a work-in-progress as part of debugging #328.  This change *should*, I would estimate, prevent the problem from being indefinite by causing a timeout and error instead.  Hopefully the diagnostic output at warn-level which outputs the last 100 trace messages will also provide enough information to understand the problem and implement a fix.

Reviewed-on: https://codeberg.org/testtrim/testtrim/pulls/331
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>