ci: fix potential deadlock on error starting TraceClient in nested-strace (#344)
Documented here: https://codeberg.org/testtrim/testtrim/issues/328#issuecomment-2664976

If `OwnedChildPid` is dropped but the child process is waiting indefinitely for the parent to send a signal to start, then the child will never terminate, causing the parent to never terminate.  The most general solution for this case is that `OwnedChildPid` should SIGTERM then SIGKILL its child, which is what is implemented here.

This affects nested strace cases only, which only applies to testtrim's own CI.

Reviewed-on: https://codeberg.org/testtrim/testtrim/pulls/344
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>