fix(strace): syscall parsing reliability improvements (#365)
Relevant improvements:
- Nested structures with `[]` and `{}` could currently have their parsing interrupted by any content within them having a `]` or `}` character -- although it's not 100% clear whether this is plausible with user-generated data and the syscalls currently monitored

This PR includes other changes which don't apply to syscalls that testtrim currently parses.  These changes were made in an effort to get testtrim to parse all syscalls in order to debug #352; supporting reading various argument types that are not apparently present in strace output for syscalls we need to parse today, but might be in the future.

re: #352 - no fixes here are relevant to fixing it, but efforts so far are looking like either missing or suppressed syscalls are still causing problems for ProcSynchronizer.
Reviewed-on: https://codeberg.org/testtrim/testtrim/pulls/365
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>