fix(strace): remove tgkill diagnostic error (#366)
In attempting to diagnose and fix #352, a new check was added in https://codeberg.org/testtrim/testtrim/pulls/361 that catches `tgkill` syscalls and errors out if the threads aren't known.  This helped to discover a problem with the process synchronizer which was fixed in https://codeberg.org/testtrim/testtrim/issues/363.

However, the diagnostic from #361 turned out to be faulty -- `tgkill` signals routinely occur *before* the strace output indicates that the threads exist at all.  Take, for example, thread `932977` in this trace:
```bash
$ rg 932977
18654:15:09:11 [ERROR] tgerror previous: "932975 tgkill(932956, 932977, SIGURG <unfinished ...>"
18661:15:09:11 [ERROR] tgerror next: "932977 --- SIGURG {si_signo=SIGURG, si_code=SI_TKILL, si_pid=932956, si_uid=1000} ---"
18693:15:09:11 [ERROR] tgerror next: "932977 clone(child_stack=0xc000118000, flags=CLONE_VM|CLONE_FS|CLONE_FILES|CLONE_SIGHAND|CLONE_THREAD|CLONE_SYSVSEM|CLONE_SETTLS <unfinished ...>"
18699:15:09:11 [ERROR] tgerror next: "932977 <... clone resumed>, tls=0xc000100098) = 932978"
18700:15:09:11 [ERROR] tgerror next: "932956 <... clone resumed>, tls=0xc0000aae98) = 932977"
19158:15:09:11 [ERROR] tgerror next: "932977 exit_group(0 <unfinished ...>"
19166:15:09:11 [ERROR] tgerror next: "932977 <... exit_group resumed>)        = ?"
19175:15:09:11 [ERROR] tgerror next: "932977 +++ exited with 0 +++"
69585:15:11:45 [ERROR] error occurred in run_tests: RunTestsErrors(UnexpectedTestError(Other(process ProcessId(932977) was supposed to be part of thread group 932956, but we did not have it tracked; could be an external process but why would testing be signaling it?
```

In this case `932977` is sent a signal before the clone returns to its parent, which triggers the "was supposed to be part of a thread group" error.  Unfortunately this seems a somewhat unavoidable race -- there's no reason why another process like `932975` can't send this signal.

This PR removes the error condition on the tgkill event.  This probably means that #352 is fixed and this diagnostic won't be helpful anymore.

Reviewed-on: https://codeberg.org/testtrim/testtrim/pulls/366
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>